+++ index.html
@@ -12,6 +12,7 @@
 <body>
   <script src="js/functions.js"></script>
   <script src="js/main.js"></script>
+  <script type="module" src="main.js"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
